+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Supported Tact 1.4.2
 
 ### Fixed
+- IRBuilder: Save trait definitions: PR [#95](https://github.com/nowarp/misti/pull/95)
 
 ## [0.1.2] - 2024-08-06