+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - IRBuilder: Mark nodes without successors as `Exit` kind: PR [#80](https://github.com/nowarp/misti/pull/80)
-- Supported Tact 1.4.2
+- Supported Tact 1.4.4
 
 ### Fixed
 - IRBuilder: Save trait definitions: PR [#95](https://github.com/nowarp/misti/pull/95)